# src/OpenMetrc.Scraper/OpenApi/metrc-swagger-v2.yaml
swagger: 2.0
info:
  title: METRC API
  version: v2
paths:
...
  /plants/v2/:
    delete:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Veg/Flower Plants • Destroy Veg/Flower Plants'
      externalDocs:
        x-doc-key: delete_plants_v2_
      operationId: DeletePlants
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility that intends to destroy the plants.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/additives:
    get:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View/Manage Plants Additives'
      externalDocs:
        x-doc-key: get_plants_v2_additives
      operationId: GetPlantAdditives
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to retrieve plant additives.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
    post:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: Manage Plants Additives'
      externalDocs:
        x-doc-key: post_plants_v2_additives
      operationId: PostPlantAdditives
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to record plant additives.
        format: ''
      responses:
        200:
          description: Ok
...
  /plants/v2/flowering:
    get:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Veg/Flower Plants'
      externalDocs:
        x-doc-key: get_plants_v2_flowering
      operationId: GetPlantFlowering
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to return the list of flowering plants.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /plants/v2/growthphases:
    get:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: <i>none</i>'
      externalDocs:
        x-doc-key: get_plants_v2_growthphases
      operationId: GetPlantGrowthPhases
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to return the list of available Plant Growth Phases.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/{id}:
    get:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Veg/Flower Plants'
      externalDocs:
        x-doc-key: get_plants_v2_{id}
      operationId: GetPlantById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        description: If specified, the Plant will be validated against the specified License Number. If not specified, the Plant will be validated against all of the User's current Facilities. Please note that if the Plant is not valid for the specified License Number, a 401 Unauthorized status will be returned.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/inactive:
    get:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Veg/Flower Plants'
      externalDocs:
        x-doc-key: get_plants_v2_inactive
      operationId: GetPlantInactive
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to return the list of inactive plants.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /plants/v2/{label}:
    get:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Veg/Flower Plants'
      externalDocs:
        x-doc-key: get_plants_v2_{label}
      operationId: GetPlantByLabel
      parameters:
      - type: string
        name: label
        in: path
        required: true
        description: Route parameter label
        format: ''
      - type: string
        name: licenseNumber
        in: query
        description: If specified, the Plant will be validated against the specified License Number. If not specified, the Plant will be validated against all of the User's current Facilities. Please note that if the Plant is not valid for the specified License Number, a 401 Unauthorized status will be returned.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/onhold:
    get:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Veg/Flower Plants'
      externalDocs:
        x-doc-key: get_plants_v2_onhold
      operationId: GetPlantOnHold
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to return the list of plants on hold.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /plants/v2/vegetative:
    get:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Veg/Flower Plants'
      externalDocs:
        x-doc-key: get_plants_v2_vegetative
      operationId: GetPlantVegetative
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to return the list of vegetating plants.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /plants/v2/waste/methods/all:
    get:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: <i>none</i>'
      externalDocs:
        x-doc-key: get_plants_v2_waste_methods_all
      operationId: GetPlantWasteMethodAll
      parameters:
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/waste/reasons:
    get:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: <i>none</i>'
      externalDocs:
        x-doc-key: get_plants_v2_waste_reasons
      operationId: GetPlantWasteReasons
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to return the list of waste reasons.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/additives/bylocation:
    post:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: Manage Plants • Manage Plants Additives'
      externalDocs:
        x-doc-key: post_plants_v2_additives_bylocation
      operationId: PostPlantAdditiveByLocation
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to record plant additives.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/manicure:
    post:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Veg/Flower Plants • Manicure/Harvest Veg/Flower Plants'
      externalDocs:
        x-doc-key: post_plants_v2_manicure
      operationId: PostPlantManicure
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to record the list of plants manicured.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/plantbatch/packages:
    post:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Immature Plants • Manage Immature Plants Inventory • View Veg/Flower Plants • Manage Veg/Flower Plants Inventory • View Packages • Create/Submit/Discontinue Packages'
      externalDocs:
        x-doc-key: post_plants_v2_plantbatch_packages
      operationId: PostPlantPlantBatchPackages
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to record the list of plant batch packages.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/plantings:
    post:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Immature Plants • Manage Immature Plants Inventory • View Veg/Flower Plants • Manage Veg/Flower Plants Inventory'
      externalDocs:
        x-doc-key: post_plants_v2_plantings
      operationId: PostPlantPlantings
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to record the list of plant batches.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/waste:
    post:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: Manage Plants Waste'
      externalDocs:
        x-doc-key: post_plants_v2_waste
      operationId: PostPlantWaste
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to record waste.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/growthphase:
    put:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Veg/Flower Plants • Manage Veg/Flower Plants Inventory'
      externalDocs:
        x-doc-key: put_plants_v2_growthphase
      operationId: PutPlantGrowthPhase
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to update the list of plants growth phase.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/harvest:
    put:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Veg/Flower Plants • Manicure/Harvest Veg/Flower Plants'
      externalDocs:
        x-doc-key: put_plants_v2_harvest
      operationId: PutPlantHarvest
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to update the list of plant harvests.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/location:
    put:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Veg/Flower Plants • Manage Veg/Flower Plants Inventory'
      externalDocs:
        x-doc-key: put_plants_v2_location
      operationId: PutPlantLocation
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to update the list of plants moved.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/merge:
    put:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Veg/Flower Plants • Manicure/Harvest Veg/Flower Plants'
      externalDocs:
        x-doc-key: put_plants_v2_merge
      operationId: PutPlantMerge
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to update the list of plants merged.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/split:
    put:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Plant'
      externalDocs:
        x-doc-key: put_plants_v2_split
      operationId: PutPlantSplit
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to update plant splits.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/strain:
    put:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Veg/Flower Plants • Manage Veg/Flower Plants Inventory'
      externalDocs:
        x-doc-key: put_plants_v2_strain
      operationId: PutPlantStrain
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to update the list of plant strains.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/tag:
    put:
      tags:
      - Plants
      summary: ''
      description: '<b>Permissions Required</b>: View Veg/Flower Plants • Manage Veg/Flower Plants Inventory'
      externalDocs:
        x-doc-key: put_plants_v2_tag
      operationId: PutPlantTag
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to upate the list of plant tags.
        format: ''
      responses:
        200:
          description: Ok
...